server/requirements.txt
asgiref==3.8.1
cffi==1.17.1
coverage==7.8.0
cryptography==44.0.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
